refactor: extract color functions into a shared module
